+++ src/main/java/com/newtongroup/library/Controller/RemoveObjectController.java
@@ -95,7 +95,7 @@ public class RemoveObjectController {
 
         for(Book temp : bookList){
             if(temp.getIsbn().equals(isbn)){
-                int id = temp.getId();
+                Long id = temp.getId();
                 String title = temp.getTitle();
                 String publisher = temp.getPublisher();
                 String price = temp.getPurchasePrice();
@@ -125,7 +125,7 @@ public class RemoveObjectController {
 
         for(EBook temp : bookList){
             if(temp.getIsbn().equals(isbn)){
-                int id = temp.getId();
+                Long id = temp.getId();
                 String title = temp.getTitle();
                 String publisher = temp.getPublisher();
                 String price = temp.getPurchasePrice();
